- Fixed typo in StastTiles.util.ts - Moved StatsTiles to StatsTiles folder
