+++ src/sections/stats/sections/overview/StatsOverview.tsx
@@ -2,7 +2,7 @@ import { useApiPromise } from "utils/api"
 import { ChartWrapper } from "./components/ChartWrapper/ChartWrapper"
 import { OmnipoolAssetsTableWrapper } from "./components/OmnipoolAssetsTable/OmnipoolAssetsTableWrapper"
 import { PieWrapper } from "./components/PieWrapper/PieWrapper"
-import { StatsTiles } from "../../components/tiles/StatsTiles"
+import { StatsTiles } from "../../components/StatsTiles/StatsTiles"
 import { isApiLoaded } from "utils/helpers"
 import { useOmnipoolOverviewData } from "./data/OmnipoolOverview.utils"
 import { useMedia } from "react-use"
